+++ packages/x-components/src/views/popular-searches.vue
@@ -55,7 +55,6 @@
   import ClearHistoryQueries from '../x-modules/history-queries/components/clear-history-queries.vue';
   import HistoryQueries from '../x-modules/history-queries/components/history-queries.vue';
   import NextQueries from '../x-modules/next-queries/components/next-queries.vue';
-  import NoSuggestions from '../x-modules/no-suggestions/components/no-suggestions.vue';
   import PopularSearches from '../x-modules/popular-searches/components/popular-searches.vue';
   import QuerySuggestion from '../x-modules/query-suggestions/components/query-suggestion.vue';
   import QuerySuggestions from '../x-modules/query-suggestions/components/query-suggestions.vue';
@@ -88,7 +87,6 @@
       HistoryQueries,
       ClearHistoryQueries,
       NextQueries,
-      NoSuggestions,
       QuerySuggestion,
       QuerySuggestions,
       RelatedTags