+++ packages/x-components/src/x-modules/no-suggestions/index.ts
@@ -1,5 +1,0 @@
-export * from './components';
-export * from './config.types';
-export * from './store';
-export * from './wiring';
-export * from './x-module';